Working test case / sketch for simple shear
The challenge here turned out to be the definition of
the analytical velocity gradient tensor used in the
manuscript. I had been assuming that c(1) > c(2) > c(3)
and transforming D onto that frame. However, the simple
shear example is set up such that the (plane strain)
problem always had c(3) == 1.0 and the rotation was
around the X(3) axis common to D on the global and
FES reference frame. This means that c(1) > c(3) > c(2)
and a general reordering by axis length transposes the
axes. For the moment insert a function to flip the
axes back again for this case.

Thus far this does not actually pass the results back
into the core of AnPAR and some cleanup is needed. But
I think this fixes the main challange. We should check
the proper ordering of axes in AnPAR (i.e. is there a
cunning switch in the Fortran / MAPLE varent we should
replecate for a MATLAB reproduction of the simple shear
and corner flow examples.

Finally, note that we will always need to special case
the first step to stick the first strain to the external
frame of reference.